Added `encoded_icon` helper
