+++ app/helpers/rails_icons/icon_helper.rb
@@ -1,0 +1,25 @@
+# frozen_string_literal: true
+
+module RailsIcons
+  module IconHelper
+    def icon(name, library: RailsIcons.configuration.default_library, from: library, variant: nil, **arguments)
+      RailsIcons::Icon.new(
+        name: name,
+        library: from || library,
+        variant: variant,
+        arguments: arguments
+      ).svg
+    end
+
+    def encoded_icon(name, library: RailsIcons.configuration.default_library, from: library, variant: nil, **arguments)
+      svg_content = RailsIcons::Icon.new(
+        name: name,
+        library: from || library,
+        variant: variant,
+        arguments: arguments
+      ).svg
+
+      "data:image/svg+xml;base64,#{Base64.strict_encode64(svg_content)}"
+    end
+  end
+end
